==> Rmd/04-qPCR-from-2022-samples.Rmd <==
---
title: "Amalga-2022-qPCR-initial-analysis"
output: html_notebook
---

30 October 2022

Cleaning up the input data to figure out the distribution of samples across qPCR plates. 
Unlike the data from 2021, the qPCR plates for the 2022 samples used a single set of standards 
rather than a full set of standards on every plate.

If there is a pattern of variance caused by a plate effect, I need to rerun the plates that 
did not have the full three replicates of 7 standards.

The metadata as originally entered was such a fiasco, I manually added additional variables (transect, depth, time-of-day) to the Excel spreadsheet to make sure nothing got horribly out of order from the code.

General background:

Sampling occurred on 3 days:
5 May 2022 (AM and PM sampling for capturing incoming and outgoing tides + vertical data using the Niskin)
9 May 2022 (sampling in the PM on the incoming tide)
25 May 2022 (sampling in the AM on the incoming tide)


That info needs to be added to the metadata.

```{r load-libraries}
library(readxl)
library(tidyverse)
library(dplyr)

```


```{r read-in-data}
qpcr.data <- read_xlsx("../data/2022_qPCR_results_combined.xlsx", sheet = "organizedData")

# metadata
meta <- read_xlsx("../data/Amalga2022_cleanMetadata.xlsx") 

```





First thing is to separate out the blanks because the syntax is all over the place (although I fixed some of it already).
```{r}
meta.blanks.explicit <- meta %>%
  mutate(sample_type = ifelse(str_detect(`Sample Code`, "Field") | str_detect(`Sample Code`, "field"), "field_blank", "sample")) %>%
    mutate(sample_type = ifelse(str_detect(`Sample Code`, "Extraction") | str_detect(`Sample Code`, "extraction"), "extraction_blank", sample_type))

```

Select just the relevant and cleaned columns from that metadata
```{r}
meta.blanks.explicit %>%
  dplyr::select(`Extraction ID`, Sample_Date, transect, distance_from_pens, depth, time_of_day, replicate, sample_type, `Sample Code`) %>% # just the relevant columns
  filter(sample_type == "sample") %>%
  group_by(Sample_Date, transect, distance_from_pens, depth, time_of_day) %>%
  tally()

# and verify that the metadata exists for all samples and 3 replicates per sample site/date
```



Clean up the metadata
```{r}
meta.organized <- meta.blanks.explicit %>%
  dplyr::select(`Extraction ID`, Sample_Date, transect, distance_from_pens, depth, time_of_day, replicate, sample_type, `Sample Code`) %>% # just the relevant columns
  mutate(tide = ifelse(Sample_Date == "2022_05_10", "incoming", NA)) %>% # create a new variable "tide" and designate incoming vs. outgoing
    mutate(tide = ifelse(Sample_Date == "2022_05_25", "incoming", tide)) %>%
      mutate(tide = ifelse(Sample_Date == "2022_05_05" & time_of_day == "AM", "outgoing", tide)) %>%
        mutate(tide = ifelse(Sample_Date == "2022_05_05" & time_of_day == "PM", "incoming", tide)) %>%
          mutate(parallel_transect = ifelse(str_detect(transect, "B2"), "-500", NA)) %>%
             mutate(parallel_transect = ifelse(str_detect(transect, "B3"), "-1000", parallel_transect)) %>%
               mutate(parallel_transect = ifelse(str_detect(transect, "T2"), "500", parallel_transect)) %>%
                 mutate(parallel_transect = ifelse(str_detect(transect, "T3"), "1000", parallel_transect)) %>%
                   mutate(parallel_transect = ifelse(str_detect(transect, "perpendicular") & distance_from_pens == 1000, "0", parallel_transect)) 


```



Some general data/metadata:
```{r}
meta.organized %>%
  group_by(Sample_Date, time_of_day, tide) %>%
  tally()

```




Check the parallel transect metadata
```{r}
meta.organized %>%
  filter(!is.na(parallel_transect)) %>%
  group_by(Sample_Date, parallel_transect) %>%
  tally()

```
```{r}
# output that metadata for porting to the VM for the current CTD analysis.
meta.organized %>%
  write_csv("csv_outputs/2022_amalga_metadata_orderly.csv")


```


## qPCR data


```{r}
head(qpcr.data)

amalga.qpcr <- qpcr.data %>%
  dplyr::select(sample, Task, CT, `Ct Mean`, `Ct SD`, Quantity, `Quantity Mean`, `Quantity SD`) 
  
```


Then stick the metadata to the qPCR data:

```{r}
# clean up the messy merged df
merged.df <-meta.organized %>%
  left_join(., amalga.qpcr, by = c("Extraction ID" = "sample")) %>%
  filter(Task != "STANDARD") 

```
Now that the metadata and qPCR data frame are cleaned-up, let's do more QC checks:


Confirm that the correct number of samples per category (depth, transect, etc.) exist.

```{r}
merged.df %>%
  group_by(Sample_Date, distance_from_pens, depth, transect, time_of_day, replicate) %>%
  tally()

```
Most samples should have 9 replicates, although I ran some on the qPCR twice. It's possible that if they had < 9, those were qPCR samples where amplification did not match the other replicates (check on this).






Then with that df, we can begin looking at the distribution of DNA quantity over the transects and depths!

```{r}
merged.df %>%
  filter(transect == "perpendicular",
         Sample_Date == "2022_05_05",
         distance_from_pens == "0",
         depth == "5"
         ) %>%
  arrange(distance_from_pens, depth) %>%
  select(`Extraction ID`, CT, Quantity, tide, time_of_day)

```

```{r}
merged.df %>%
  filter(`Extraction ID` %in% c("e01291", "e01293", "e01294", "e01295")) %>%
  select(`Extraction ID`, distance_from_pens, CT)


```

```{r}
merged.df %>%
  arrange(`Extraction ID`) %>%
  select(`Extraction ID`, distance_from_pens, depth, CT, Quantity)

```
How far out CT-value-wise do the standards go?
```{r}
qpcr.data %>%
  filter(Task == "STANDARD", sample == "standard6")

```

standard 7 is mostly undetermined, but pops up around 37.5-40.1 cycles
standard 6 is also sometimes undetermined, but pops up at 34.5-36.5 cycles

seems like actual quantification drops off after 36ish cycles.

What are the CT values for the samples?
```{r}
qpcr.data %>%
  filter(Task == "UNKNOWN", CT != "Undetermined") %>%
  arrange(CT) %>%
  filter(CT > 37)

```
samples:
CT = 25-44.5 cycles.
How many samples would be excluded if we dropped anything >37?

A lot! 189 rows!

better to use the gBlock to push the quantification out farther (closer to 40). Or just go with the results as they are and change later if necessary.


```{r checking-data-for-inhibition}
merged.df %>%
  filter(distance_from_pens == "0" &
           CT != "Undetermined" &
           depth == "0") %>%
  select(`Extraction ID`, transect, distance_from_pens, depth, time_of_day, tide, CT, Quantity, Sample_Date) %>%
  ggplot(aes(x = `Extraction ID`, y = CT, color = Sample_Date)) +
  geom_point()
  
```



# Plot data

```{r vertical-transect}
# first look at the vertical transect data
merged.df %>%
  filter(transect == "perpendicular",
         Sample_Date == "2022_05_05"
         ) %>%
  arrange(distance_from_pens) %>%
    mutate(tide = ifelse(tide == "outgoing", "outgoing_AM", "incoming_PM")) %>%
  ggplot(aes(x = distance_from_pens, y = log10(`Quantity Mean`), color = tide)) +
  geom_point(size = 3, alpha = .3) +
  facet_grid(rows = vars(depth), labeller = label_both) +
  theme_bw() +
  labs(
    y = "DNA mean copy number (log10)",
    x = "Distance from hatchery pens (m)",
    title = "Perpendicular transect; vertical sampling (surface-10 m)",
    color = "Tide"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/amalgaMay2022_verticalDepthTransect.pdf", height = 6, width = 8)
```

Ok, but that uses the qPCR mean quantity, rather than calculating the mean from the replicates.

Look at the individual variation across qPCR replicates:

```{r}
# first look at the vertical transect data
merged.df %>%
  filter(transect == "perpendicular",
         Sample_Date == "2022_05_05"
         ) %>%
  arrange(distance_from_pens) %>%
    mutate(tide = ifelse(tide == "outgoing", "outgoing_AM", "incoming_PM")) %>%
  ggplot(aes(x = distance_from_pens, y = log10(Quantity), color = tide)) +
  geom_point(size = 3, alpha = .3) +
  facet_grid(rows = vars(depth), labeller = label_both) +
  theme_bw() +
  labs(
    y = "DNA copy number (log10)",
    x = "Distance from hatchery pens (m)",
    title = "Perpendicular transect; vertical sampling (surface-10 m)",
    color = "Tide"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/amalgaMay2022_verticalDepthTransect_qPCR_reps.pdf", height = 6, width = 8)
```




## Parallel transect

Take a look at the transect parallel to the hatchery pens; perpendicular to the original transect:



```{r}
# make the parallel transect variable numeric
merged.df$parallel_transect <- as.numeric(merged.df$parallel_transect)

merged.df %>%
  filter(str_detect(transect, "parallel") | transect == "perpendicular" & distance_from_pens == 1000,
         Sample_Date == "2022_05_05",
         !is.na(depth)
         ) %>%
  arrange(parallel_transect) %>%
    mutate(tide = ifelse(tide == "outgoing", "outgoing_AM", "incoming_PM")) %>%
  ggplot(aes(x = parallel_transect, y = log10(`Quantity Mean`), color = tide)) +
  geom_point(size = 3, alpha = 0.3) +
  facet_grid(rows = vars(depth), labeller = label_both) +
  theme_bw() +
  labs(
    y = "DNA mean copy number (log10)",
    x = "surface transect (m), SE to NW, parallel to pens",
    title = "Vertical depth transect (surface to 10 m) parallel to pens"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))


```
Interesting. What about NOT log-scaled?

```{r}
merged.df %>%
  filter(str_detect(transect, "parallel")  | distance_from_pens == 1000,
         Sample_Date == "2022_05_05",
         !is.na(depth)
         ) %>%
  arrange(parallel_transect) %>%
  mutate(tide = ifelse(tide == "outgoing", "outgoing_AM", "incoming_PM")) %>%
  ggplot(aes(x = parallel_transect, y = `Quantity Mean`, color = tide)) +
  geom_point(size = 3, alpha = 0.3) +
  facet_grid(rows = vars(depth), labeller = label_both) +
  theme_bw() +
  labs(
    y = "DNA mean copy number",
    x = "surface transect (m), SE to NW, parallel to pens",
    title = "Vertical depth transect (surface to 10 m) parallel to pens"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +
  scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/parallel_transect_w_depth.pdf", width = 8, height = 6)

```

One observation: within 500m has more DNA than within 1000m in either direction, with what appears to be a bias toward more chum DNA on the incoming tide (?) and to the southeast of the net pens when compared to the equivalent distances to the northwest of the pens. In other words, if water movement in Lynn Canal is from the NW toward the SE, then this would correspond with what we're seeing in chum DNA movement.


```{r}
merged.df

```



What about looking at surface flow from the other days?

```{r}
merged.df %>%
  filter(str_detect(transect, "parallel")  | distance_from_pens == 1000,
         !is.na(depth),
         depth == 0
         ) %>%
  arrange(parallel_transect) %>%
    mutate(tide = ifelse(tide == "outgoing", "outgoing_AM", "incoming_PM")) %>%
  ggplot(aes(x = parallel_transect, y = `Quantity Mean`, color = tide)) +
  geom_point(size = 3, alpha = 0.3) +
  facet_grid(rows = vars(Sample_Date)) +
  theme_bw() +
  labs(
    y = "DNA mean copy number",
    x = "surface transect (m), SE to NW, parallel to pens",
    title = "Surface transect parallel to pens"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/parallel_transect_w_days.pdf", width = 8, height = 6)
```
Think about wind moving surface water, especially on 5/10/2022.

Wind direction in the AM vs. PM. Basically surface questions so worth looking at winds more than tides?
5/5/2022 wind from W/SW in the morning (basically calm in the morning)
5/5/2022 afternoon wind from the north.


With all qPCR reps:
```{r}
merged.df %>%
  filter(str_detect(transect, "parallel")  | distance_from_pens == 1000,
         !is.na(depth),
         depth == 0
         ) %>%
  arrange(parallel_transect) %>%
    mutate(tide = ifelse(tide == "outgoing", "outgoing_AM", "incoming_PM")) %>%
  ggplot(aes(x = parallel_transect, y = Quantity, color = tide)) +
  geom_point(size = 3, alpha = 0.3) +
  facet_grid(rows = vars(Sample_Date), scales = "free_y") +
  theme_bw() +
  labs(
    y = "DNA copy number",
    x = "surface transect (m), SE to NW, parallel to pens",
    title = "Surface transect parallel to pens"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/parallel_transect_w_days_qPCRreps.pdf", width = 8, height = 6)
```


Normal transect across multiple days:

```{r}
# first look at the vertical transect data
merged.df %>%
  filter(transect == "perpendicular",
         tide == "incoming") %>%
  arrange(distance_from_pens) %>%
  ggplot(aes(x = distance_from_pens, y = log10(`Quantity Mean`), color = tide)) +
  geom_point(size = 3, alpha = .3) +
  facet_grid(rows = vars(Sample_Date)) +
  theme_bw() +
  labs(
    y = "DNA mean copy number (log10)",
    x = "Distance from hatchery pens (m)",
    title = "Perpendicular transect; surface samples",
    color = "Tide"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/amalgaMay2022_perpendicularSurfaceTransect.pdf", height = 6, width = 8)
```

I wonder if there's inhibition in any of the samples from 5/25? Strange to see such a high drop-out incidence, but also notable because May was such a dry month and there was a ton of pollen present on the water surface.


```{r}
# first look at the vertical transect data
merged.df %>%
  filter(transect == "perpendicular",
         tide == "incoming") %>%
  arrange(distance_from_pens) %>%
  ggplot(aes(x = distance_from_pens, y = log10(Quantity), color = tide)) +
  geom_point(size = 3, alpha = .3) +
  facet_grid(rows = vars(Sample_Date)) +
  theme_bw() +
  labs(
    y = "DNA copy number (log10)",
    x = "Distance from hatchery pens (m)",
    title = "Perpendicular transect; surface samples",
    color = "Tide"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) +   scale_color_manual(values = c("goldenrod", "dodgerblue"))

ggsave("pdf_outputs/amalgaMay2022_perpendicularSurfaceTransect_qPCRreps.pdf", height = 6, width = 8)
```


## qPCR QC


I need to dig more into drop-outs, etc. in the qPCR data.

```{r}
# first check: qPCR replicates for the same extraction
merged.df$CT <- as.numeric(merged.df$CT)

merged.df %>%
  mutate(CT_r = round(CT, digits = 3)) %>%
  group_by(`Extraction ID`) %>%
  ggplot(aes(x = log10(Quantity), y = CT_r)) +
  geom_point()

```
That relationship between log10 DNA copy number and CT is spot-on!

```{r}
merged.df %>%
  filter(is.na(CT) & !is.na(`Ct Mean`),
         `Ct Mean` < 37) # none of the drop-outs have a CT value < 37
```
 Seems likely that replicates with a Ct Mean > 40 are prone to drop-out at a higher rate.
 
 
The plot supposedly has 432 rows removed - that seems like a higher number than the number of NAs in the CT column...
```{r}
merged.df %>%
  filter(is.na(CT))
```
432/735 are NAs!

```{r}
# all replicates are NAs?
merged.df %>%
  filter(is.na(CT) & is.na(`Ct Mean`))
  
```
338 rows for which all three reps are NAs

How many are from 2000 m in the perpendicular transect or 1000 m in the parallel transect?
```{r}
merged.df %>%
  filter(is.na(CT) & is.na(`Ct Mean`),
         distance_from_pens == 2000 | parallel_transect %in% c(1000, -1000))


```

For those samples/replicates with more than 1 successful replicate, what were the `Ct SD` values?
```{r}
merged.df %>%
  filter(!is.na(`Ct Mean`)) %>%
  arrange(desc(`Ct SD`)) %>%
  mutate(prop_SD = `Ct SD`/`Ct Mean`) %>%
  filter(prop_SD <0.5)

```
282/397 rows = the Ct SD < 5% of the Ct Mean.




## Working in > 2D?


Try to make this 3D

http://htmlpreview.github.io/?https://github.com/AckerDWM/gg3D/blob/master/gg3D-vignette.html

```{r}
#devtools::install_github("AckerDWM/gg3D")
library(gg3D)

```

To use these data in 3D space, I need to adjust the y-axis identity for the parallel transect data points (e.g., all were taken at y=1000m)

```{r}
df.for.3d <- merged.df %>%
  mutate(distance_from_pens = ifelse(!is.na(parallel_transect), 1000, distance_from_pens))


df.for.3d %>%
  filter(str_detect(transect, "parallel"),
         Sample_Date == "2022_05_05",
         !is.na(depth)
         ) %>%
  arrange(parallel_transect) %>%
  ggplot(aes(x = distance_from_pens, y = log10(`Quantity Mean`), z = depth, color = tide)) +
  geom_point(size = 3)+
  axes_3D() +
  stat_3D()


  +
  theme_bw() +
  labs(
    y = "DNA mean copy number (log10)",
    x = "Distance from hatchery pens (m)",
    title = "Vertical depth transect (surface-10 m)",
    color = "Tide"
  ) +
  theme(
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(margin = margin(r = 10))
  ) 
```



### Wind data from AOOS


```{r}

wind <- read_csv("../data/org_mxak_little_island_9636_c4d0_84e2_AOOSwindData.csv", col_names = T, skip = 1)
head(wind)

# need to convert the time from UTC to 
Sys.timezone()

# Alaska daylight time is UTC -8 hours
ak.wind <- with_tz(wind, tzone = "America/Anchorage") %>%
  rename(ak.time = UTC)

```

```{r}
may.wind <- ak.wind %>%
  filter(month(ak.time)==05 & day(ak.time) %in% c(05,10,25)) %>%
  select(ak.time, `m.s-1`, degrees)

```
Note: Wind direction is measured in degrees clockwise from due north. Hence, a wind coming from the south has a wind direction of 180 degrees; one from the east is 90 degrees.


```{r}
# plot the wind for 5-10-2021
may.wind %>%
  filter(year(ak.time)==2021 & day(ak.time)==10) %>%
  mutate(time_of_day = ifelse(hour(ak.time)>11, "PM", "AM")) %>%
  ggplot(aes(x = degrees, y = `m.s-1`, fill = time_of_day)) +
  geom_col(width = 20, alpha = 0.5)  + 
  scale_x_continuous(breaks = seq(0, 359, by = 10), limits = c(-10, 350)) +
  coord_polar(start = -pi/18, clip = "off") +
  theme_bw() +
  labs(
    y = "wind speed m*s-1",
    x = "wind direction",
    title = "10 May 2021, Little Island, SE AK",
    fill = "time of day"
  ) +
  scale_fill_manual(values = c("goldenrod", "midnightblue"))
  
ggsave("pdf_outputs/amalgaMay2021_windplot.pdf")
```


Make the plots for 2022
Probably worth constraining these time periods to the actual sampling times

```{r}
# May 5, 2022
may5 <- may.wind %>%
  filter(year(ak.time)==2022 & day(ak.time)==5 & hour(ak.time)>7 & hour(ak.time)<17) %>%
  mutate(time_of_day = ifelse(hour(ak.time)>11, "PM", "AM")) %>%
  ggplot(aes(x = degrees, y = `m.s-1`, fill = time_of_day)) +
  geom_col(width = 20, alpha = 0.3)  + 
  scale_x_continuous(breaks = seq(0, 359, by = 10), limits = c(-10, 350)) +
  coord_polar(start = -pi/18, clip = "off") +
  facet_grid(cols = vars(time_of_day))+
  theme_bw() +
  labs(
    y = "wind speed m*s-1",
    x = "wind direction",
    title = "5 May 2022",
    fill = "time of day"
  ) +
  scale_fill_manual(values = c("goldenrod", "midnightblue")) +
  theme(
  strip.background = element_blank(),
  strip.text.x = element_blank()
  )

may5
```


```{r}
# May 10, 2022
may10 <- may.wind %>%
  filter(year(ak.time)==2022 & day(ak.time)==10) %>%
  mutate(time_of_day = ifelse(hour(ak.time)>11, "PM", "AM")) %>%
  filter(time_of_day == "PM") %>%
  ggplot(aes(x = degrees, y = `m.s-1`, fill = time_of_day)) +
  geom_col(width = 20, alpha = 0.3)  + 
  scale_x_continuous(breaks = seq(0, 359, by = 10), limits = c(-10, 350)) +
  coord_polar(start = -pi/18, clip = "off") +
  theme_bw() +
  labs(
    y = "wind speed m*s-1",
    x = "wind direction",
    title = "10 May 2022",
    fill = "time of day"
  ) +
  scale_fill_manual(values = c("midnightblue"))  +
  theme(
    legend.position = "none"
  )

may10
```



```{r}
# May 25, 2022
may25 <- may.wind %>%
  filter(year(ak.time)==2022 & day(ak.time)==25) %>%
  mutate(time_of_day = ifelse(hour(ak.time)>11, "PM", "AM")) %>%
  filter(time_of_day == "AM") %>%
  ggplot(aes(x = degrees, y = `m.s-1`, fill = time_of_day)) +
  geom_col(width = 20, alpha = 0.3)  + 
  scale_x_continuous(breaks = seq(0, 359, by = 10), limits = c(-10, 350)) +
  coord_polar(start = -pi/18, clip = "off") +
  theme_bw() +
  labs(
    y = "wind speed m*s-1",
    x = "wind direction",
    title = "25 May 2022",
    #fill = "time of day"
  ) +
  scale_fill_manual(values = c("goldenrod", "midnightblue")) +
  theme(
    legend.position = "none"
  )

may25
```


```{r}
library(cowplot)

p2 <- plot_grid(may10, may25, align = "h", axis = "bt")

plot_grid(may5, p2, ncol = 1, align = "v", axis = 'l')

ggsave("pdf_outputs/amalgaMay2022_comboWindPlot.pdf", width = 8, height = 9)
```
