==> Rmd/08-ctd-cast-data.Rmd <==
---
title: "08-ctd-cast-data"
output: html_notebook
---

12 February 2022

CTD casts from eDNA sampling off the Sashin on 10 May, 2021.

```{r load-libraries}
library(tidyr)
library(readr)
library(stringr)
library(lubridate)
library(hms)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(lubridateExtras)

```

```{r read-in-csv-files}
# read in all files in the reheaded directory
# dataFiles <- lapply(Sys.glob("CTDCasts/reheaded/*.csv"), read_csv)
# 
# # turn each of the data frames in that list into a single data frame
# ctd.data <- do.call("rbind", dataFiles)

ctd.w.tides.dist <- read_csv("../data/ctdDataframe.csv") %>%
  filter(distance != 100)


```


```{r data-clean-up}
# organize the headers, etc.
# ctd.data2 <- ctd.data %>%
#   rename(pressure_decibar = `Pressure (Decibar)`, depth_m = `Depth (Meter)`, temp_c = `Temperature (Celsius)`, conductivity = `Conductivity (MicroSiemens per Centimeter)`, sp_conduct = `Specific conductance (MicroSiemens per Centimeter)`, salinity = `Salinity (Practical Salinity Scale)`, sound_velocity = `Sound velocity (Meters per Second)`, density = `Density (Kilograms per Cubic Meter)`, file = `% File name`, lat = `% Start latitude`, long = `% Start longitude`, duration = `% Cast duration (Seconds)`, time = `% Cast time (local)`) %>%
#   separate(file, into = c(NA, NA, "ctd_sample"), sep = "_") %>%
#   select(ctd_sample, time, lat, long, depth_m, salinity, density, pressure_decibar, temp_c, conductivity, sp_conduct, sound_velocity, duration)
# 
# 
# # add tide info
# ctd.w.tides <- ctd.data2 %>%
#   mutate(tide = ifelse(hour(hms::as_hms(time)) >12, "PM_outgoing", "AM_incoming"))
# 
# ctd.w.tides %>%
#   select(lat) %>%
#   unique() %>%
#   arrange(lat)

```



## Convert longitude to meters from chum pens

This will match the other plots in the series.
```{r}
# meta <- read_csv("../metadata/amalga_clean_metadata.csv")
# 
# dist <- meta %>%
#   filter(!is.na(distance)) %>%
#   arrange(distance) %>%
#   select(distance) %>%
#   unique()
# 
# dist2x <- dist %>%
#   bind_rows(dist)

# # each cast (ctd_sample) represents one of the 80m sampling intervals
# ctd.w.dist <- ctd.w.tides %>%
#   arrange(time) %>%
#   select(time) %>%
#   unique() %>%
#   mutate(id = row_number()) %>%
#   bind_cols(dist2x)
# 
# # bind it back together
# ctd.w.tides.dist <- ctd.w.tides %>%
#   left_join(., ctd.w.dist)
  
```



```{r}
# QC: are the data what we think they are?
ctd.w.tides.dist %>%
    mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  ggplot(aes(x = long, y = lat, color = tide, linetype = tide, label = ctd_sample)) +
 # geom_line() +
  #geom_point(size = 0.1)+
  geom_text(size = 2.5)+
  facet_grid(rows = vars(tide)) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))
  ) +
  labs(y = "Latitude",
       x = "Longitude") +
  scale_color_manual(values = c("firebrick1", "midnightblue"))

ggsave("pdf_outputs/ctdSimpleLatLon.pdf")

```



Output that dataframe so that I can plot the transect on the bathy map
```{r}
# ctd.w.tides.dist %>%
#   write_csv("csv_outputs/ctdDataframe.csv")

```



Ultimately, turn this into a map of the transect rather than an x-y plot.

max/min surface water temperature vs. at depth
```{r summary-water-temps}
ctd.w.tides.dist %>%
  #filter(distance == 0) %>%
  ggplot(aes(x = temp_c, y = -1*depth_m, color = tide)) +
  geom_point()


```
```{r temp-at-10m-depth}
ctd.w.tides.dist %>%
  filter(depth_m > 9 & depth_m < 11) %>%
  summarise(min(temp_c))


```

```{r temp-at-surface}
ctd.w.tides.dist %>%
  group_by(tide) %>%
  filter(depth_m > 1) %>%
  summarise(min(temp_c))



```




## Plotting CTD profiles

```{r}
# salinity
ctd.w.tides.dist %>%
  filter(salinity > 20) %>%
  mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  ggplot(aes(x = density, y = -1*(depth_m), color = distance)) +
  geom_point() +
  facet_grid(rows = vars(tide)) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))
  ) +
  labs(y = "Depth (m)",
       x = "Density",
       color = "Distance")
  

#ggsave("pdf_outputs/ctdSalinityProfile.pdf", width = 8, height = 5)
```
Salinity: zoomed in
```{r}
# temp
ctd.w.tides.dist %>%
    filter(salinity > 20) %>%
  mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 10) %>% # top three meters of the surface
  ggplot(aes(x = density, y = -1*(depth_m), color = distance)) +
  geom_point() +
  facet_grid(rows = vars(tide)) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))) +
  labs(y = "Depth (m)",
       x = "Density (kg/m^3)",
       color = "Distance (m)")

#ggsave("pdf_outputs/ctdSalinityProfile5m.pdf", width = 8, height = 5)

```

# Plotting a surface with temperature and longitude

https://github.com/cathmmitchell/plottingOceanDataWithR/wiki/Irregular-data

```{r}
binned <- ctd.w.tides.dist %>% 
    mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 10) %>%
  mutate_at(5, round)%>% 
  group_by(distance, tide, depth_m) %>% 
  summarize(av_tmp = mean(temp_c))

# flip order so higher temps are at the top!
binned

tempplot <- ggplot(binned,aes(x=distance,y=-1*(depth_m))) +
  facet_grid(rows = vars(tide))+
  geom_contour_filled(aes(z=av_tmp), alpha = 0.8) +
    geom_point(size =0.1) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))) +
  labs(y = "Depth (m)",
       x = "Distance from hatchery pens (m)",
       fill = "Temperature (C)") +
  scale_x_continuous(expand = c(0,0)) +
  scale_y_continuous(expand = c(0,0.05)) +
  scale_fill_brewer(palette = "Spectral", direction = -1,
                    guide = guide_legend(reverse = T))
  

tempplot

ggsave("pdf_outputs/ctdTemp10metersDistance.png", width = 8, height = 5)
```
Why do some of the top depth values drop out? Something about the binning.


Do it once more, but setting the depth even shallower:

```{r temp-binned-2}
binned2 <- ctd.w.tides.dist %>% 
    mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 2.5) %>%
  mutate_at(5, round, 2)%>% 
  group_by(distance, tide, depth_m) %>% 
  summarize(av_tmp = mean(temp_c))


ggplot(binned2,aes(x=distance,y=-1*(depth_m))) +
  facet_grid(rows = vars(tide))+
  geom_contour_filled(aes(z=av_tmp), alpha = 0.8) +
  geom_point(size =0.1) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))) +
  labs(y = "Depth (m)",
       x = "Distance from hatchery pens (m)",
       fill = "Temperature (C)")

ggsave("pdf_outputs/ctdTemp3metersDistance.pdf", width = 8, height = 5)
```


For salinity and temp, let's zoom in on the top 5 meters (we took samples from the first ~1 m)
```{r}
# temp
ctd.w.tides.dist %>%
  mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 5) %>%
  ggplot(aes(x = temp_c, y = -1*(depth_m), color = distance)) +
  geom_point() +
  facet_grid(rows = vars(tide)) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))) +
  labs(y = "Depth (m)",
       x = "Temperature (C)",
       color = "Distance (m)")

ggsave("pdf_outputs/ctdTempProfile3metersDistance.pdf", width = 8, height = 5)

```

### Plotting a surface with salinity and longitude


```{r salinity-plot}

binnedSalinity <- ctd.w.tides.dist %>% 
    mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 10) %>%
  mutate_at(5, round)%>% 
  group_by(distance, tide, depth_m) %>% 
  summarize(av_salinity = mean(salinity))

# oranize colors
mycolors <- colorRampPalette(brewer.pal(9, "Oranges"))(11)


salplot <- ggplot(binnedSalinity, aes(x = distance, y = -1*(depth_m))) +
  facet_grid(rows = vars(tide)) +
  geom_contour_filled(aes(z = av_salinity), alpha = 0.8) +
    geom_point(size = 0.1) +
    theme_bw() +
     theme(
      axis.title.x = element_text(margin = margin(t=10)),
      axis.title.y = element_text(margin = margin(r=10))) +
      labs(y = "Depth (m)",
           x = "Distance from hatchery pens (m)",
           fill = "Salinity (PSU)") +
  scale_x_continuous(expand = c(0,0)) +
  scale_y_continuous(expand = c(0,0.05))


salplot

ggsave("pdf_outputs/ctdSalinity10metersDistance.png", width = 8, height = 5)
```

```{r salinity-binned-2}
binnedSal2 <- ctd.w.tides.dist %>% 
    mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 2.5) %>%
  mutate_at(5, round, 2)%>% 
  group_by(distance, tide, depth_m) %>% 
  summarize(av_sal = mean(salinity))


ggplot(binnedSal2,aes(x=distance,y=-1*(depth_m))) +
  facet_grid(rows = vars(tide))+
  geom_contour_filled(aes(z=av_sal), alpha = 0.8) +
  geom_point(size =0.1) +
  theme_bw()+
   theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10))) +
  labs(y = "Depth (m)",
       x = "Distance from hatchery pens (m)",
       fill = "Salinity (PSU)")

ggsave("pdf_outputs/ctdSalinity3metersDistance.png", width = 8, height = 5)
```


```{r make-SI-figure-S2}
library(patchwork)

salplot + tempplot + plot_layout(nrow = 2) + plot_annotation(tag_levels = "A")

ggsave("pdf_outputs/ctd_data2021_dualplot.png", width = 8, height = 8)
```


```{r}
ctd.w.tides.dist %>%
  group_by(tide) %>%
  summarise(max(salinity))
```




Density

```{r}
density1 <- ctd.w.tides.dist %>% 
  mutate(tide = ifelse(tide == "AM_incoming", "incoming tide (AM)", "outgoing tide (PM)")) %>%
  filter(depth_m < 10) %>%
  mutate_at(5, round)%>% 
  group_by(distance, tide, depth_m) %>% 
  summarize(av_density = mean(density)) # the avg needs to be taken for proper plotting in this way.

# ggplot(density1,aes(x=distance,y=-1*(depth_m))) +
#   facet_grid(rows = vars(tide))+
#   geom_contour_filled(aes(z=density), alpha = 0.8) +
#     geom_point(size =0.1) +
#   theme_bw()+
#    theme(
#     axis.title.x = element_text(margin = margin(t=10)),
#     axis.title.y = element_text(margin = margin(r=10))) +
#   labs(y = "Depth (m)",
#        x = "Distance from hatchery pens (m)",
#        fill = "Density (kg/m^3)")

#ggsave("pdf_outputs/ctdTemp10metersDistance.pdf", width = 8, height = 5)
```



## Tide data

Tide data was obtained from the NOAA Tides website: tidesandcurrents.noaa.gov

https://tidesandcurrents.noaa.gov/waterlevels.html?id=9452210&units=standard&bdate=20210510&edate=20210511&timezone=LST/LDT&datum=MLLW&interval=6&action=
```{r read-in-tide-data}
tides <- read_csv("../data/JuneauTides_2021May.csv",
                  col_types = cols_only(Date=col_date("%Y/%m/%d"),
                              `Time (LST/LDT)`=col_time("%H:%M"),
                              `Verified (ft)`=col_double())) 
# fix variable names
tides.renamed <- tides %>%
  rename(time = `Time (LST/LDT)`, height_ft = `Verified (ft)`)

# create a merged column with date and time to then round to the hour as in the CTD data
tides.renamed$newdate <- with(tides.renamed, as.POSIXct(paste(Date, time), format="%Y-%m-%d %H:%M:%S"))

tides.hour <- tides.renamed %>%
  mutate(round_time = round_date(newdate, unit = "hours")) %>% # round time stamps to the hour
  separate(round_time, into = c("ymd", "hour"), sep = " ") %>%
  mutate(hour = as_hms(hour)) 
```
Look at joining the tide data with the ctd data based on time

```{r plot-tidal-cycle-2021}
# install.packages("devtools")
#library(devtools)
#devtools::install_github("ellisvalentiner/lubridateExtras")

# from the ctd time stamp, I need to separate the date from the time
ctd.w.tides.dist$time <- ymd_hms(ctd.w.tides.dist$time)


ctd.times <- ctd.w.tides.dist %>%
  mutate(round_time = round_date(time, unit = "hours")) %>% # round time stamps to the hour
  separate(round_time, into = c("ymd", "hour"), sep = " ") %>%
  mutate(hour = as_hms(hour)) 
  

tidePlot2021 <- tides.hour %>%
  filter(Date == "2021-05-10" | Date == "2021-05-09") %>%
  left_join(., ctd.times, by = c("hour", "ymd")) %>%
  ggplot(aes(x=newdate, y=height_ft, color = distance)) +
  geom_point() +
  theme_bw() +
  theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10)),
    legend.title = element_text(margin = margin(b=10))) +
  labs(y = "Tide height (ft)",
       x = "Date-time",
       color = "Distance from hatchery pens (m)") +
  scale_color_viridis_c(option = "plasma", direction = 1)
  
tidePlot2021

ggsave("pdf_outputs/tideCycleSampling.png", width = 8, height = 5)

```

```{r alternative-plot-2021}
# removing-distance-and-just-using-sampling-period

# tidePlot2021 <- tides.hour %>%
#   filter(Date == "2021-05-10" | Date == "2021-05-09") %>%
#   left_join(., ctd.times, by = c("hour", "ymd")) %>%
#   mutate(sampling_period = ifelse(is.na(ctd_sample), "no", "yes")) %>%
#   ggplot(aes(x = newdate, y = height_ft, color = sampling_period)) +
#   geom_point() +
#   theme_bw() +
#   theme(
#     axis.title.x = element_text(margin = margin(t=10)),
#     axis.title.y = element_text(margin = margin(r=10)),
#     legend.title = element_text(margin = margin(b=10))) +
#   labs(y = "Tide height (ft)",
#        x = "Date-time",
#        color = "Sampling period") +
#   #scale_color_viridis_c(option = "plasma", direction = -1) +
# scale_color_manual(values = c("gray", "darkgreen"))
#  
# tidePlot2021


```




```{r import-tide-data-2022}
# read in tide data from 2022
tides2022 <- read_csv("../data/JuneauTides_2022May.csv",
                  col_types = cols_only(Date=col_date("%Y/%m/%d"),
                              `Time (LST/LDT)`=col_time("%H:%M"),
                              `Verified (ft)`=col_double())) 
# fix variable names
tides.renamed2 <- tides2022 %>%
  rename(time = `Time (LST/LDT)`, height_ft = `Verified (ft)`)

# create a merged column with date and time to then round to the hour as in the CTD data
tides.renamed2$newdate <- with(tides.renamed2, as.POSIXct(paste(Date, time), format="%Y-%m-%d %H:%M:%S"))

tides.hour2 <- tides.renamed2 %>%
  mutate(round_time = round_date(newdate, unit = "hours")) %>% # round time stamps to the hour
  separate(round_time, into = c("ymd", "hour"), sep = " ") %>%
  mutate(hour = as_hms(hour))
```

```{r CTD-data-2022}
# import CTD data for May-5
ctd.2022 <- read_csv("../data/2022ctdDataframe.csv")

# grab the data for May 5 and one tide for data points for the map
one.set.2022 <- ctd.2022 %>%
  filter(str_detect(time, "2022-05-05")) %>%
  #dplyr::select(time, ctd_sample, lat, long) %>%
  unique() %>%
  filter(!ctd_sample %in% c("163024", "191935", "175337", "185047")) %>%
  mutate(Year = "2022")

one.set.2022

# from the ctd time stamp, I need to separate the date from the time
one.set.2022$time <- ymd_hms(one.set.2022$time)

ctd.times.2022 <- one.set.2022 %>%
  mutate(round_time = round_date(time, unit = "hours")) %>% # round time stamps to the hour
  separate(round_time, into = c("ymd", "hour"), sep = " ") %>%
  mutate(hour = as_hms(hour))
  
# read in metadata for distance from pens associated with ctd times
dist_time2022 <- read_csv("csv_outputs/2022_ctd_grp_metadata.csv") %>%
  dplyr::select(ctd_sample, tide, distance_from_pens, sample_code) %>%
  unique()

# add distance to ctd_sample number
ctd.times.2022 <- ctd.times.2022 %>%
  left_join(., dist_time2022, by = "ctd_sample")
```


```{r plot-tidal-cycle-2022}
tidePlot2022 <- tides.hour2 %>%
  filter(Date == "2022-05-04" | Date == "2022-05-05") %>%
  left_join(., ctd.times.2022, by = c("hour", "ymd")) %>%
  mutate(sampling_period = ifelse(is.na(ctd_sample), "no", "yes")) %>%
  ggplot(aes(x = newdate, y = height_ft, color = sampling_period)) +
  geom_point() +
  theme_bw() +
  theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10)),
    legend.title = element_text(margin = margin(b=10))) +
  labs(y = "Tide height (ft)",
       x = "Date-time",
       color = "Sampling period") +
  #scale_color_viridis_c(option = "plasma", direction = -1) +
scale_color_manual(values = c("gray", "darkgreen"))

  
tidePlot2022

ggsave("pdf_outputs/tideCycleSampling2022.png", width = 8, height = 5)

```


```{r mapping-distance-from-pens-onto-tides}
tides.hour2 %>%
  filter(Date == "2022-05-04" | Date == "2022-05-05") %>%
  left_join(., ctd.times.2022, by = c("hour", "ymd")) %>%
  mutate(sampling_period = ifelse(is.na(ctd_sample), "no", "yes")) %>%
  ggplot(aes(x = newdate, y = height_ft, color = distance_from_pens)) +
  geom_point() +
  theme_bw() +
  theme(
    axis.title.x = element_text(margin = margin(t=10)),
    axis.title.y = element_text(margin = margin(r=10)),
    legend.title = element_text(margin = margin(b=10))) +
  labs(y = "Tide height (ft)",
       x = "Date-time",
       color = "Distance from hatchery pens (m)") +
   scale_color_viridis_c(option = "plasma", direction = 1)
  
```



```{r}
# double check the sampling to see if it overlaps slack tide?
tides.hour2 %>%
  filter(Date == "2022-05-05") %>%
  left_join(., ctd.times.2022, by = c("hour", "ymd")) %>%
  filter(tide == "AM_outgoing") %>%
  mutate(sampling_period = ifelse(is.na(ctd_sample), "no", "yes")) %>%
  filter(height_ft < min(height_ft)+1) %>%
  ggplot(aes(x = long, y = lat, color = newdate)) +
  geom_point()

```
Minimum tidal height was at 10:36; the parallel transect was sampled after 11 am, so it was as the tide turned.



```{r combine-plots-both-years}
# make supplemental figure
tidePlot2021 + tidePlot2022 + 
  plot_layout(ncol = 1) +
  plot_annotation(tag_levels = "A")

ggsave("pdf_outputs/SIfigure_tideCycle_bothYears_v2.png", width = 8, height = 8)

```
